+++ rust/worker/src/system/mod.rs
@@ -1,10 +1,12 @@
 mod executor;
+mod receiver;
 mod scheduler;
-mod sender;
 mod system;
 mod types;
+mod wrapped_message;
 
 // Re-export types
-pub(crate) use sender::*;
+pub(crate) use receiver::*;
 pub(crate) use system::*;
 pub(crate) use types::*;
+pub(crate) use wrapped_message::*;
